requirements.txt
stdiomask
pyperclip
pyqt6
pycryptodome
argon2-cffi
zxcvbn
requests